+++ cpp/src/distance/neighbors/specializations/ivfflat_extend_uint8_t_int64_t.cu
@@ -29,8 +29,7 @@ namespace raft::neighbors::ivf_flat {
                        raft::neighbors::ivf_flat::index<T, IdxT>* idx,                        \
                        raft::device_matrix_view<const T, IdxT, row_major> new_vectors,        \
                        std::optional<raft::device_vector_view<const IdxT, IdxT>> new_indices);
-
-RAFT_MAKE_INSTANCE(float, uint64_t);
+RAFT_MAKE_INSTANCE(uint8_t, int64_t);
 
 #undef RAFT_MAKE_INSTANCE
